+++ Modules/Microsoft365DSC/DSCResources/MSFT_SPOApp/MSFT_SPOApp.psm1
@@ -64,10 +64,6 @@ function Get-TargetResource
 
     Write-Verbose -Message "Getting configuration for app $Identity"
 
-    # Temp workaround for Graph connection issue. Make sure connecting to Graph first (#xxxx)
-    $ConnectionModeGraph = New-M365DSCConnection -Workload 'MicrosoftGraph' `
-        -InboundParameters $PSBoundParameters
-
     $ConnectionMode = New-M365DSCConnection -Workload 'PnP' `
         -InboundParameters $PSBoundParameters
 
@@ -202,10 +198,6 @@ function Set-TargetResource
         -Parameters $PSBoundParameters
     Add-M365DSCTelemetryEvent -Data $data
     #endregion
-
-    # Temp workaround for Graph connection issue. Make sure connecting to Graph first (#xxxx)
-    $ConnectionModeGraph = New-M365DSCConnection -Workload 'MicrosoftGraph' `
-        -InboundParameters $PSBoundParameters
 
     $ConnectionMode = New-M365DSCConnection -Workload 'PnP' `
         -InboundParameters $PSBoundParameters
@@ -366,10 +358,6 @@ function Export-TargetResource
 
     try
     {
-        # Temp workaround for Graph connection issue. Make sure connecting to Graph first (#xxxx)
-        $ConnectionModeGraph = New-M365DSCConnection -Workload 'MicrosoftGraph' `
-            -InboundParameters $PSBoundParameters
-
         $ConnectionMode = New-M365DSCConnection -Workload 'PnP' `
             -InboundParameters $PSBoundParameters