+++ Modules/Microsoft365DSC/DSCResources/MSFT_SPOBrowserIdleSignout/MSFT_SPOBrowserIdleSignout.psm1
@@ -62,10 +62,6 @@ function Get-TargetResource
 
     Write-Verbose -Message 'Getting configuration for SPO Browser Idle Signout settings'
 
-    # Temp workaround for Graph connection issue. Make sure connecting to Graph first (#xxxx)
-    $ConnectionModeGraph = New-M365DSCConnection -Workload 'MicrosoftGraph' `
-        -InboundParameters $PSBoundParameters
-
     $ConnectionMode = New-M365DSCConnection -Workload 'PnP' `
         -InboundParameters $PSBoundParameters
 
@@ -196,10 +192,6 @@ function Set-TargetResource
         -Parameters $PSBoundParameters
     Add-M365DSCTelemetryEvent -Data $data
     #endregion
-
-    # Temp workaround for Graph connection issue. Make sure connecting to Graph first (#xxxx)
-    $ConnectionModeGraph = New-M365DSCConnection -Workload 'MicrosoftGraph' `
-        -InboundParameters $PSBoundParameters
 
     $ConnectionMode = New-M365DSCConnection -Workload 'PnP' `
         -InboundParameters $PSBoundParameters
@@ -358,10 +350,6 @@ function Export-TargetResource
 
     try
     {
-        # Temp workaround for Graph connection issue. Make sure connecting to Graph first (#xxxx)
-        $ConnectionModeGraph = New-M365DSCConnection -Workload 'MicrosoftGraph' `
-            -InboundParameters $PSBoundParameters
-
         $ConnectionMode = New-M365DSCConnection -Workload 'PNP' `
             -InboundParameters $PSBoundParameters