+++ Modules/Microsoft365DSC/DSCResources/MSFT_SPOHomeSite/MSFT_SPOHomeSite.psm1
@@ -56,10 +56,6 @@ function Get-TargetResource
     )
     Write-Verbose -Message "Getting configuration for hub site collection $Url"
 
-    # Temp workaround for Graph connection issue. Make sure connecting to Graph first (#xxxx)
-    $ConnectionModeGraph = New-M365DSCConnection -Workload 'MicrosoftGraph' `
-        -InboundParameters $PSBoundParameters
-
     $ConnectionMode = New-M365DSCConnection -Workload 'PnP' `
         -InboundParameters $PSBoundParameters
 
@@ -186,10 +182,6 @@ function Set-TargetResource
     Add-M365DSCTelemetryEvent -Data $data
     #endregion
     Write-Verbose -Message "Setting configuration for home site '$Url'"
-
-    # Temp workaround for Graph connection issue. Make sure connecting to Graph first (#xxxx)
-    $ConnectionModeGraph = New-M365DSCConnection -Workload 'MicrosoftGraph' `
-        -InboundParameters $PSBoundParameters
 
     $ConnectionMode = New-M365DSCConnection -Workload 'PnP' `
         -InboundParameters $PSBoundParameters