+++ Modules/Microsoft365DSC/DSCResources/MSFT_SPOHubSite/MSFT_SPOHubSite.psm1
@@ -614,10 +614,6 @@ function Export-TargetResource
 
     try
     {
-        # Temp workaround for Graph connection issue. Make sure connecting to Graph first (#xxxx)
-        $ConnectionModeGraph = New-M365DSCConnection -Workload 'MicrosoftGraph' `
-            -InboundParameters $PSBoundParameters
-
         $ConnectionMode = New-M365DSCConnection -Workload 'PnP' `
             -InboundParameters $PSBoundParameters