+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change log for Microsoft365DSC
 
+# Unreleased
+
+* OD and SPO resources
+  * Implemented PnP/Graph conflict issue workaround (Issue 4746)
+
 # 1.24.619.1
 
 * SCDLPComplianceRule